chore: move conditions utilities to dedicated file
